SDK: the Node plugin file sources move out of the inspector server
`createDirectoryPluginFileSource`, `createZipPluginFileSource` and
`collectZipArchiveObservations` lived in the inspector server, which was fine
while the server was their only consumer. The CLI's local package modes need
exactly these three, and both alternatives were bad: import them from the
inspector server (a CLI depending on a web server) or write a second copy —
two implementations of the archive rules the OpenAI package lane grades, which
is the mirror-the-SDK mistake this codebase has made once already, and whose
symptom is two readings of one ZIP quietly disagreeing.

They are on the SDK's NODE entry only. `plugin-bundle/index.ts` stays free of
`node:fs` and of an archive library, because that freedom is what lets a
browser validate a dropped package in the page — and a re-export from the pure
barrel would end it without anyone noticing until a bundle broke. The
browser-entry guard still reports an empty set of Node builtins.

The server keeps a re-export shim so its five importers compile unchanged. The
archive-observation test travelled with its implementation: a test left behind
would be guarding a shim rather than the rules it is about.
