+++ sdk/src/index.ts
@@ -533,6 +533,17 @@ export {
 // pure data and data reasoning only, so importing the result model or the
 // package reader never pulls a transport in with it.
 export * from "./openai-readiness/index.js";
+// The Node plugin-bundle file sources: a directory on disk and a ZIP in
+// memory. NODE ENTRY ONLY — they are the only `plugin-bundle` modules that
+// touch `node:fs` or an archive library, and `plugin-bundle/index.ts` stays
+// free of both so a browser can still validate a dropped package in the page.
+export {
+  DIRECTORY_ARCHIVE_OBSERVATIONS,
+  collectZipArchiveObservations,
+  createDirectoryPluginFileSource,
+  createZipPluginFileSource,
+} from "./plugin-bundle/node-file-sources.js";
+
 // The Node XML parser for SVG dimension reads, exported ONLY here. A browser
 // has `DOMParser` natively and `readImageDimensions` finds it; `@xmldom/xmldom`
 // is banned from the browser entry's import graph, so the Node fallback lives
